backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
